pyproject.toml
[tool.poetry]
name = "python-fastapi-openapi-k8s"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.119.1"
gunicorn = "^23.0.0"
uvicorn = "^0.38.0"
pytz = "^2025.2"
httpx = "^0.28.1"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
